screenshot: Select a queue with both present and graphics capabilities
Keep a list of all queues in the device map structure so as to be
able to use an appropriate queue when taking a screenshot. (The queue
passed in to vkQueuePresent may not have the VK_QUEUE_GRAPHICS
capability set, so it can't be used for doing a blit.) The list
of available queues is searched for a queue that is graphics
and present capable, and that is the queue that is used for
getting the screenshot.

Since a queue is a dispatchable object, added code to
create a dispatch table each queue.

Change-Id: I18c2767811a37e4c3ce59ddda98201631cbde875